+++ tests/unit_tests/i15_1/conftest.py
@@ -1,0 +1,51 @@
+from pathlib import Path
+
+import pytest
+from dodal.devices.beamlines.i15_1.robot import Robot
+from dodal.devices.tetramm import TetrammDetector
+from dodal.devices.zebra.zebra import Zebra, ZebraMapping
+from ophyd_async.core import StaticFilenameProvider, StaticPathProvider, init_devices
+from ophyd_async.epics.motor import Motor
+
+
+@pytest.fixture
+def path_provider() -> StaticPathProvider:
+    return StaticPathProvider(StaticFilenameProvider(""), Path(""))
+
+
+@pytest.fixture
+async def i0(path_provider: StaticPathProvider) -> TetrammDetector:
+    async with init_devices(mock=True):
+        i0 = TetrammDetector(
+            "",
+            path_provider,
+            name="i0",
+        )
+    return i0
+
+
+@pytest.fixture
+async def zebra() -> Zebra:
+    async with init_devices(mock=True):
+        zebra = Zebra(
+            ZebraMapping(),
+            "",
+            "zebra",
+        )
+    return zebra
+
+
+@pytest.fixture
+async def robot() -> Robot:
+    async with init_devices(mock=True):
+        robot = Robot("", "")
+
+    return robot
+
+
+@pytest.fixture
+async def tth() -> Motor:
+    async with init_devices(mock=True):
+        tth = Motor("", "")
+
+    return tth
