+++ tests/unit_tests/i15_1/test_take_eiger_data.py
@@ -1,20 +1,25 @@
-from pathlib import Path
-
 import pytest
 from bluesky.simulators import RunEngineSimulator, assert_message_and_return_remaining
-from ophyd_async.core import StaticFilenameProvider, StaticPathProvider, init_devices
+from dodal.devices.beamlines.i15_1.robot import Robot
+from dodal.devices.tetramm import TetrammDetector
+from dodal.devices.zebra.zebra import Zebra
+from ophyd_async.core import StaticPathProvider, init_devices
+from ophyd_async.epics.motor import Motor
 from ophyd_async.fastcs.eiger import EigerDetector
 
-from crystallography_bluesky.i15_1.plans.take_eiger_data import take_eiger_data
+from crystallography_bluesky.i15_1.plans.take_eiger_data import (
+    take_eiger_and_i0_data,
+    take_eiger_data,
+)
 
 
 @pytest.fixture
-async def eiger() -> EigerDetector:
+async def eiger(path_provider: StaticPathProvider) -> EigerDetector:
     async with init_devices(mock=True):
         eiger = EigerDetector(
             name="fastcs-eiger",
             prefix="",
-            path_provider=StaticPathProvider(StaticFilenameProvider(""), Path("")),
+            path_provider=path_provider,
         )
     return eiger
 
@@ -82,9 +87,13 @@ def test_take_eiger_data_makes_expected_calls(eiger: EigerDetector):
     )
 
 
-def test_take_eiger_and_i0_data_makes_expected_calls(eiger: EigerDetector):
+def test_take_eiger_and_i0_data_makes_expected_calls(
+    eiger: EigerDetector, i0: TetrammDetector, zebra: Zebra, robot: Robot, tth: Motor
+):
     run_engine = RunEngineSimulator()
-    msgs = run_engine.simulate_plan(take_eiger_data(10, 0.1, eiger))
+    msgs = run_engine.simulate_plan(
+        take_eiger_and_i0_data(10, 0.01, eiger, i0, zebra, robot, tth)
+    )
     msgs = assert_message_and_return_remaining(
         msgs,
         predicate=lambda msg: (
@@ -102,11 +111,6 @@ def test_take_eiger_and_i0_data_makes_expected_calls(eiger: EigerDetector):
     msgs = assert_message_and_return_remaining(
         msgs,
         predicate=lambda msg: msg.command == "stage" and msg.obj.name == "i0",
-    )
-
-    msgs = assert_message_and_return_remaining(
-        msgs,
-        predicate=lambda msg: msg.command == "wait",
     )
     msgs = assert_message_and_return_remaining(
         msgs,
@@ -174,11 +178,11 @@ def test_take_eiger_and_i0_data_makes_expected_calls(eiger: EigerDetector):
     )
     msgs = assert_message_and_return_remaining(
         msgs,
-        predicate=lambda msg: (
-            msg.command == "unstage" and msg.obj.name == "fastcs-eiger"
-        ),
+        predicate=lambda msg: msg.command == "unstage" and msg.obj.name == "i0",
     )
     msgs = assert_message_and_return_remaining(
         msgs,
-        predicate=lambda msg: msg.command == "unstage" and msg.obj.name == "i0",
+        predicate=lambda msg: (
+            msg.command == "unstage" and msg.obj.name == "fastcs-eiger"
+        ),
     )
